requirements.txt
langchain
langchain-openai
langchain-google-genai 
python-dotenv
markdown-it-py
PyMuPDF
streamlit
streamlit_pdf_viewer